app/requirements.txt
dj-database-url==2.0.0
Django==3.2.25
django-bootstrap-v5==1.0.0
mysqlclient==2.2.0
python-decouple==3.8
# Pillow
psycopg2-binary>=2.8
Pillow==9.0.0

#intelx
requests

#SPF
dnspython

#Shodan
shodan

#Criptography
cryptography